Вынес редактирование в общий компонент
